Updated manuscript
I have removed the 'no pmm' part from the Heckman model. Additionally, I've included the plot function with echo=FALSE, making it suitable for copying and pasting into the appendix later on. However, the pattern function with cluster still does not work. Regarding the box 3 about the prediction matrix, previously i have added relevant information about this (with an equation) in the Imputation flow. Please review this information to ensure that the box is constructed appropriately, avoiding redundant information throughout the paper. Thanks.